Serialize most f64 parameters as xsd:double
Many of the f64 parameters in our struct definitions are defined as being of
type xsd:double in the XML Schema for DASH. The Rust serialization for these
does not correspond to xsd:double for INF, -INF and NaN, so we use a dedicated
serialization function.

Bug reported by @ypo. Fixes issue #49.