Downloading: attempt to fix A/V desynchronization when muxing.
When the audio and video tracks are unsynchronized due to a difference in their
startTime attribute, we attempt to fix this desynchronization during muxing.
This is a rare problem in the wild and has not been heavily tested. The fix is
currently only implemented when using ffmpeg as a muxer application (uses the
`-itsoffset` commandline option).