Significant changes to MPD serialization for better specification conformity.
In particular, the order of elements in the Rust structures used for serialization and
deserialization of an MPD now follows that of the XML schema for DASH. Please see the CHANGELOG.md
for more details of the associated changes, which justify a version bump.

This will be released as a beta version to allow further testing by users.