Tests: fix rewrite-rickroll test.
We need to try harder to delete the audio AdaptationSet: the contentType or mimeType attribute may
be set on the underlying Representation(s), rather than on the AdaptationSet element directly.

Recent versions of ffmpeg complain if an empty stream is provided to the -map commandline argument
(which was the case for this test, because the audio AdaptationSet was not correctly deleted by the
XSLT stylesheet, and its content was being replaced by a segment containing only video.