Preliminary support for security sandboxing on Linux using Landlock.
This allows for restricting privileges available to our application code and our helper applications
on Linux, using the Landlock loadable security module. This functionality is gated by the `sandbox`
crate feature, and also currently needs to be enabled at runtime by calling the `sandbox` method on
`DashDownloader`. The feature restricts access to the filesystem, limiting write access to
directories that we will need to write to, limiting exec access to directories where our helper
applications and their runtime libraries are located, and limiting read access to directories that
the application or helper applications may need to read. It also somewhat restricts network access,
preventing binding to a TCP port.